+++ src/constants/plugins.ts
@@ -22,7 +22,7 @@ import $12 from '@plugins/message'
 import $13 from '@plugins/memberships'
 import $14 from '@kazu80/clubs-plugin-links'
 import $15 from '@plugins/join-legacy'
-import $16 from '@plugins/veritrans'
+import $16 from '@plugins/clubs-payments'
 import $17 from '@plugins/tickets'
 
 import * as PostsPlugin from '@devprotocol/clubs-plugin-posts'
