+++ src/plugins/clubs-payments/slot-currency-option.svelte
