+++ src/plugins/clubs-payments/slot.astro
